Several fixes
The reason for the container to not be able to access /dev/kvm
is that, in opensuse, the kvm group is 492. This patch fixes
this by changing the group to kvm before creating the containers
and restoring it after. It's better than setting 0666
permissions.

Also, the 3 seconds delay to generate the fake key
authentication not always does work, so this patch also adds a
loop to retry until it suceeds.